backend/requirements.txt
# ============================================================================
# RiceGuard Backend Requirements (FastAPI + MongoDB + ML)
# ============================================================================

# ----------------------------------------------------------------------------
# Core web stack
# ----------------------------------------------------------------------------
fastapi==0.103.2
uvicorn[standard]==0.23.2
starlette==0.27.0

# ----------------------------------------------------------------------------
# Validation & request parsing
# ----------------------------------------------------------------------------
pydantic[email]==1.10.19
email-validator==2.1.0.post1
python-multipart==0.0.9

# ----------------------------------------------------------------------------
# Configuration / environment handling
# ----------------------------------------------------------------------------
python-dotenv==1.0.1

# ----------------------------------------------------------------------------
# Database (MongoDB Atlas compatible)
# ----------------------------------------------------------------------------
pymongo[srv]==4.6.3    # "[srv]" pulls in dnspython for mongodb+srv URIs

# ----------------------------------------------------------------------------
# Security / authentication
# ----------------------------------------------------------------------------
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# ----------------------------------------------------------------------------
# Machine learning / image processing (optional on Python < 3.13)
# ----------------------------------------------------------------------------
tensorflow==2.20.0 ; python_version < "3.13"
numpy==1.26.4 ; python_version < "3.13"
numpy>=2.1,<3.0 ; python_version >= "3.13"
pillow==10.4.0

# ----------------------------------------------------------------------------
# Observability / utilities
# ----------------------------------------------------------------------------
requests==2.31.0
typing-extensions>=4.7.1
pymongo-stubs
